chore: move requests.js ./requests/requests.js
